+++ bpf/wordcloud/execsnoop_example.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+sudo ./execsnoop.py  -T constitution.txt
+