+++ bpf/wordcloud/installdeps.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+set -e
+sudo dnf install python3-wordcloud